+++ BILBO-firmware.X/bilbo/low_level/user_leds.h
@@ -17,6 +17,6 @@ void handle_tuning_output_leds(int8_t level);
 
 void handle_range_leds_out(uint8_t range);
 
-int led_init();
+uint8_t led_init();
 
 #endif
